Dice Sets, Dice minimums, Optimizations, and tweaks
* Added dice sets
* Added dice minimums
* Added dice class default minimum = 1
* Added dice instances can be set to a specific minimum

The minimums are so that the dice don't roll into a negative number or
0, unless the user desires it to (in which case set it to nil)

* Optimization:  many of the default library functions (ie. math.random,
math.max) were set as locals to speed things up
* Changed dice:roll() no longer returns a table!  Instead it return
value(s) depending on how many dice sets there are.  If you want tables
just use {Dice:roll()} to put the results in a table
* Changed string.match for the dice string to check for a SPECIFIC
pattern.  If dice does not match pattern EXACTLY it will cause an error.

* Removed dice.odds

Calculating the odds with rerolls requires some kind of formula that I
couldn't figure out so I was brute forcing the combinations to get the
%'s.  With a big enough dice and rerolls this would have been slow as a
slug.